refactor: rename autumn.types -> autumn.models
